Add explicit platform data for EEPROM to speed it up
By default page_size is 1 for the at24 without it being
explicitly specified.  Since only page_size bytes is written
at a time, this meant to fill the EEPROM 8192 writes were needed,
and with a 1ms timeout this ends up taking a long time.
Increasing the page size to the max helps a lot.

Signed-off-by: Sarah Newman <snewman@shotspotter.com>