Fix slave audio device
- Configure CODEC as a bit and frame slave. Bitclock and Frame clock
are coming from the master CODEC on scepter. This avoids the
WM8737 driving the clock lines.
- Force the McBSP to always be a slave regardless of the Bitclock/Frame
clock settings. This is a quick hack as ALSA assumes if the CODEC is
slave, the McBSP must be master
- Force the McBSP to always be a slave regardless of the Bitclock/Frame
clock settings. This is a quick hack as ALSA assumes if the CODEC is
slave, the McBSP must be master. For the scepter case, clocks are always
coming from outside of the SoC.
- Correct DMA channel settings in the HACK for McBSP2 (1 based).
- Clean up debug/dead code

Signed-off-by: Hunyue Yau <hy-git@hy-research.com>